PATCH: [ 1431131 ] Fix VS.NET 2003 compilation   Src\Common: coretools.cpp scbarg.cpp scbarg.h sizecbar.cpp sizecbar.h

+++ Src/Common/coretools.cpp
@@ -651,9 +651,9 @@ CString LegalizeFileName(LPCTSTR szFileName)
 	return CString(tempname);
 }
 
-static double tenpow(double expon)
+static double tenpow(int expon)
 {
-	int base=10;
+	double base=10;
 	double rtn = pow(base, expon);
 	return rtn;
 }
